+++ mobile/android/android-components/components/concept/sync/src/main/java/mozilla/components/concept/sync/Devices.kt
@@ -158,6 +158,7 @@ data class DeviceConfig(
  */
 enum class DeviceCapability {
     SEND_TAB,
+    CLOSE_TABS,
 }
 
 /**